Fix global-buffer-overflow
In `struct PELE`, because of we storing and casting different pointer
types in `int *valp`, then the Address Sanitizer will report error when
we try to dereference it.

To solve this problem, we declare void pointer instead of int pointer
`valp` and aslo declare `int valsize` to store actually size of object.

For generic pointer casting, we provide `get_plist_valp()` to casting
different pointer types according to `valsize`.